requirements.txt
kivymd
kivy
plyer
pillow